optimze read by using file handle to read in iterator
