+++ be/src/olap/rowset/segment_v2/segment_iterator.h
@@ -30,6 +30,7 @@
 #include "olap/rowset/segment_v2/column_zone_map.h"
 #include "olap/rowset/segment_v2/ordinal_page_index.h"
 #include "olap/olap_cond.h"
+#include "util/file_cache.h"
 
 namespace doris {
 
@@ -120,6 +121,9 @@ private:
     // used to binary search the rowid for a given key
     // only used in `_get_row_ranges_by_keys`
     std::unique_ptr<RowBlockV2> _seek_block;
+
+    // Handle for file to read
+    std::unique_ptr<OpenedFileHandle<RandomAccessFile>> _file;
 };
 
 }
